requirements.txt
aiofiles==24.1.0
aiohttp==3.8.6
aiosignal==1.3.1
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.8
configparser==7.1.0
dataclasses-json==0.6.7
distro==1.9.0
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.2
fastapi==0.97.0
filelock==3.15.4
frontend==0.0.3
frozenlist==1.3.3
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
importlib-metadata==6.7.0
importlib_resources==6.4.3
isodate==0.6.1
itsdangerous==2.2.0
jiter==0.5.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.34
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langsmith==0.1.101
looseversion==1.3.0
lxml==5.3.0
marshmallow==3.22.0
minio==7.2.7
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.1
nibabel==5.2.1
nipype==1.8.6
numpy==1.22.4
olefile==0.47
openai==1.42.0
openpyxl==3.1.3
orjson==3.10.7
packaging==24.0
pandas==2.0.3
pathlib==1.0.1
pip==24.2
prov==2.0.1
psycopg2-binary==2.9.9
pycparser==2.21
pycryptodome==3.20.0
pyctuator==0.18.1
pydantic==1.10.17
pydot==3.0.1
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-multipart==0.0.8
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
redis==5.0.8
regex==2024.7.24
requests==2.31.0
scikit-learn==1.3.2
scipy==1.10.1
setuptools==72.1.0
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.27.0
tenacity==8.5.0
threadpoolctl==3.1.0
tiktoken==0.7.0
tqdm==4.66.5
traits==6.3.2
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
urllib3==2.0.7
uvicorn==0.22.0
wheel==0.43.0
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.15.0
